requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
chromadb==0.4.22
sentence-transformers==2.3.1
bm25s==0.1.10
python-frontmatter==1.0.1
python-dotenv==1.0.1
aiofiles==23.2.1
watchdog==4.0.0
httpx>=0.27.0
numpy>=1.24.0,<2.0
